requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0